services/asr/requirements.txt
fastapi
uvicorn
prometheus_client
torch
openai-whisper
python-multipart
numpy
ffmpeg-python
